requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
psycopg[binary]==3.3.2
python-dotenv==1.2.1
